refactor(cli): centralize model picker access view
