+++ src/test-kernel/cli/ModelListForm.vitest.mts
@@ -1,11 +1,6 @@
 import { describe, expect, it } from 'vitest';
 
 import {
-  emptyModelListMessageForCliMode,
-  formatModelStatusForCliMode,
-  modelSelectItemsForCliMode,
-} from '@cli/chat/tui/modelAccessDisplay';
-import {
   modelListDescription,
   modelSelectWindow,
 } from '@cli/chat/tui/forms/ModelListForm';
@@ -22,256 +17,6 @@ import {
   selectVisibleInlineOverflowText,
   visibleSelectRange,
 } from '@cli/chat/tui/ui/Select';
-import type { CliModelAccess } from '@cli/runtime/modelAccess';
-import type { ModelOptionData } from '@shared/schemas';
-
-function access(
-  model: Partial<ModelOptionData>,
-  status = 'api key set',
-  available = !model.disabled && !model.requiresKey,
-): CliModelAccess {
-  return {
-    model: {
-      value: 'deepseekT',
-      label: 'DeepSeek V4 Flash (Thinking)',
-      ...model,
-    },
-    available,
-    status,
-  };
-}
-
-describe('CLI ModelListForm status text', () => {
-  it('marks personal API mode explicitly', () => {
-    expect(
-      formatModelStatusForCliMode(
-        access({ availability: 'provider-key' }),
-        'personal',
-      ),
-    ).toBe('api: api key set');
-  });
-
-  it('marks included relay mode explicitly', () => {
-    expect(
-      formatModelStatusForCliMode(
-        access({ availability: 'included-access' }, 'included access'),
-        'included',
-      ),
-    ).toBe('relay: included');
-  });
-
-  it('does not describe personal-key access as relay availability', () => {
-    expect(
-      formatModelStatusForCliMode(
-        access({ availability: 'provider-key' }),
-        'included',
-      ),
-    ).toBe('relay: unavailable; api key set');
-  });
-
-  it('distinguishes exhausted relay quota from tier exclusion', () => {
-    expect(
-      formatModelStatusForCliMode(
-        access(
-          {
-            availability: 'relay-quota-exhausted',
-            availabilityLabel: 'Relay quota exhausted',
-            disabled: true,
-          },
-          'relay quota exhausted',
-        ),
-        'included',
-      ),
-    ).toBe('relay: quota exhausted');
-  });
-
-  it('shows signed-out included mode as a relay login requirement', () => {
-    expect(
-      formatModelStatusForCliMode(
-        access(
-          {
-            availability: 'included-login-required',
-            availabilityLabel: 'Login required',
-            disabled: true,
-          },
-          'login required',
-        ),
-        'included',
-      ),
-    ).toBe('relay: login required');
-  });
-});
-
-describe('CLI ModelListForm model visibility', () => {
-  it('shows only models runnable in the active API mode', () => {
-    const items = modelSelectItemsForCliMode(
-      [
-        access({
-          value: 'sonnet46T',
-          label: 'Sonnet',
-          availability: 'included-access',
-        }),
-        access(
-          {
-            value: 'deepseekT',
-            label: 'DeepSeek',
-            availability: 'provider-key',
-          },
-          'api key set',
-        ),
-        access(
-          {
-            value: 'openrouterOnlyT',
-            label: 'OpenRouter Only',
-            availability: 'openrouter-key',
-          },
-          'openrouter key',
-        ),
-        access(
-          {
-            value: 'opus48T',
-            label: 'Opus',
-            availability: 'not-included',
-            disabled: true,
-          },
-          'not included',
-          false,
-        ),
-        access({
-          value: 'gpt54',
-          label: 'GPT-5.4',
-          availability: 'included-access',
-        }),
-      ],
-      'included',
-    );
-
-    expect(items.map((item) => item.value)).toEqual(['sonnet46T', 'gpt54']);
-    expect(items.map((item) => item.description)).toEqual([
-      'relay: included',
-      'relay: included',
-    ]);
-  });
-
-  it('shows personal-key models for personal API mode', () => {
-    const items = modelSelectItemsForCliMode(
-      [
-        access(
-          {
-            value: 'sonnet46T',
-            label: 'Sonnet',
-            availability: 'included-access',
-          },
-          'included access',
-        ),
-        access({
-          value: 'deepseekT',
-          label: 'DeepSeek',
-          availability: 'provider-key',
-        }),
-        access(
-          {
-            value: 'openrouterOnlyT',
-            label: 'OpenRouter Only',
-            availability: 'openrouter-key',
-          },
-          'openrouter key',
-        ),
-        access(
-          {
-            value: 'gemini31p',
-            label: 'Gemini',
-            availability: 'missing-key',
-            requiresKey: true,
-          },
-          'missing api key',
-          false,
-        ),
-      ],
-      'personal',
-    );
-
-    expect(items.map((item) => item.value)).toEqual([
-      'deepseekT',
-      'openrouterOnlyT',
-    ]);
-    expect(items.map((item) => item.description)).toEqual([
-      'api: api key set',
-      'api: openrouter key',
-    ]);
-  });
-
-  it('marks API-mode-runnable models disabled when the live chat cannot switch formats', () => {
-    const items = modelSelectItemsForCliMode(
-      [
-        access({
-          value: 'sonnet46T',
-          label: 'Sonnet',
-          availability: 'provider-key',
-        }),
-        access({
-          value: 'gpt55',
-          label: 'GPT-5.5',
-          availability: 'provider-key',
-        }),
-      ],
-      'personal',
-      (model) =>
-        model === 'sonnet46T'
-          ? 'different conversation format; start new chat'
-          : undefined,
-    );
-
-    expect(items).toEqual([
-      {
-        value: 'sonnet46T',
-        label: 'Sonnet',
-        description:
-          'api: api key set; different conversation format; start new chat',
-        disabled: true,
-      },
-      {
-        value: 'gpt55',
-        label: 'GPT-5.5',
-        description: 'api: api key set',
-        disabled: false,
-      },
-    ]);
-  });
-
-  it('treats filtered-empty lists as non-actionable', () => {
-    expect(
-      modelSelectItemsForCliMode(
-        [
-          access(
-            {
-              availability: 'provider-key',
-            },
-            'api key set',
-            false,
-          ),
-          access(
-            {
-              availability: 'missing-key',
-              requiresKey: true,
-            },
-            'missing api key',
-            false,
-          ),
-          access(
-            {
-              availability: 'not-included',
-              disabled: true,
-            },
-            'not included',
-            false,
-          ),
-        ],
-        'included',
-      ).length,
-    ).toBe(0);
-  });
-});
 
 describe('CLI ModelListForm empty state', () => {
   it('uses a truthful description for empty and non-empty model lists', () => {
@@ -286,63 +31,6 @@ describe('CLI ModelListForm empty state', () => {
     );
     expect(modelListDescription({ itemCount: 2, selectable: true })).toBe(
       'Choose the model for future turns.',
-    );
-  });
-
-  it('explains that included relay models require sign-in', () => {
-    const models = [
-      access(
-        {
-          availability: 'included-login-required',
-          disabled: true,
-        },
-        'login required',
-        false,
-      ),
-    ];
-
-    expect(emptyModelListMessageForCliMode(models, 'included')).toBe(
-      'Included relay models require sign-in. Run /login or switch with /api personal.',
-    );
-  });
-
-  it('points included relay users at personal keys when no relay models are runnable', () => {
-    expect(
-      emptyModelListMessageForCliMode(
-        [
-          access(
-            {
-              availability: 'not-included',
-              disabled: true,
-            },
-            'not included',
-            false,
-          ),
-        ],
-        'included',
-      ),
-    ).toBe(
-      'No included relay models are runnable. Switch with /api personal or try again later.',
-    );
-  });
-
-  it('points personal API-key users at key setup or included relay', () => {
-    expect(
-      emptyModelListMessageForCliMode(
-        [
-          access(
-            {
-              availability: 'missing-key',
-              requiresKey: true,
-            },
-            'missing api key',
-            false,
-          ),
-        ],
-        'personal',
-      ),
-    ).toBe(
-      'No personal API-key models are runnable. Configure a provider key or switch with /api included.',
     );
   });
 });
